refactor : Split CalendarColorMap
